requirements.txt
flask
pymongo
python-dotenv
pymupdf
#azure-ai-openai
azure-ai-inference
azure-core
pytest
pytest-mock
azure-storage-blob